sqlite - move code from `api.py` to `__init__.py`
